Permission Validation: Extended QueryProcessor to include Provenance information in QueryResult. Used this provenance information to verify the user permissions: In all services the processing is encapsulated in the processQuery method, that checks if all required permissions are met.
Required changes: Added all group permissions to the user's permissions.
Added loadGroup method to the UserDB. Added new
PermissionDeniedException. Extended the ProvenanceService, as it now
requires additional information like the type of the result, bbox and
time. Also adusted the sourcelist request in the User service to only
return sources the user is allowed to see. All new users are now
automatically joined to the "users" group for easier permission
distribtuion on "public" data sets.